#7 | code refactor
