+++ app/middleware/authMiddleware.js
@@ -1,6 +1,6 @@
 const jwt = require('jsonwebtoken');
 const secret = process.env.SECRET;
-const { response } = require('../helpers/send-response');
+const { handleError } = require('../helpers/handle-error');
 
 module.exports = (req, res, next) => {
     if(req.method === "OPTIONS"){
@@ -9,13 +9,13 @@ module.exports = (req, res, next) => {
     try {
         const token = req.headers.authorization.split(' ')[1];
         if(!token) {
-            return response(res, 403, "User is not authorized");
+            return handleError(res, 403, "User is not authorized");
         }
         const decodedData = jwt.verify(token, secret);
         req.user = decodedData;
         next()// call next middleware
         
     } catch (error) {
-        response(res, 403, "User is not authorized");
+        handleError(res, 403, "User is not authorized");
     }
 }
